+++ tests/graphics/geometry_test.c
@@ -107,7 +107,7 @@ PalBool geometryTest()
     }
 
     // enumerate all available adapters
-    int32_t adapterCount = 0;
+    uint32_t adapterCount = 0;
     result = palEnumerateAdapters(&adapterCount, nullptr);
     if (result != PAL_RESULT_SUCCESS) {
         logResult(result, "Failed to get adapters");
@@ -346,7 +346,7 @@ PalBool geometryTest()
     }
 
     // create shaders
-    uint64_t bytecodeSize = 0;
+    uint32_t bytecodeSize = 0;
     void* bytecode = nullptr;
     const char* sources[3];
     PalShaderEntryInfo entries[3];
